+++ services/server/src/server/services/storageServices/EtherscanVerifyApiService.ts
@@ -23,7 +23,7 @@ const DEFAULT_ETHERSCAN_CHAINLIST_ENDPOINT =
   "https://api.etherscan.io/v2/chainlist";
 const DEFAULT_BLOCKSCOUT_CHAINLIST_ENDPOINT =
   "https://chains.blockscout.com/api/chains";
-const BLOCKSCOUT_ALREADY_VERIFIED = "BLOCKSCOUT_ALREADY_VERIFIED";
+const VERIFIER_ALREADY_VERIFIED = "VERIFIER_ALREADY_VERIFIED";
 const ROUTESCAN_CHAINLIST_ENDPOINTS = [
   {
     workspace: "mainnet",
@@ -324,18 +324,27 @@ export const buildJobExternalVerificationsObject = (
     }
 
     let statusUrl;
+    let contractApiUrl;
     if (
       verifierData.verificationId &&
       // We need to handle the special case for a blockscout already verified contract
-      verifierData.verificationId !== BLOCKSCOUT_ALREADY_VERIFIED
+      verifierData.verificationId !== VERIFIER_ALREADY_VERIFIED
     ) {
       try {
-        const apiBaseUrl = verifierService?.getApiUrl(
+        const statusApiBaseUrl = verifierService?.getApiUrl(
           "checkverifystatus",
           parseInt(chainId),
         );
-        if (apiBaseUrl) {
-          statusUrl = `${apiBaseUrl}&guid=${encodeURIComponent(verifierData.verificationId)}`;
+        if (statusApiBaseUrl) {
+          statusUrl = `${statusApiBaseUrl}&guid=${encodeURIComponent(verifierData.verificationId)}`;
+        }
+
+        const contractApiBaseUrl = verifierService?.getApiUrl(
+          "getabi",
+          parseInt(chainId),
+        );
+        if (contractApiBaseUrl) {
+          contractApiUrl = `${contractApiBaseUrl}&address=${address}`;
         }
       } catch (error) {
         // Cannot generate url
@@ -362,6 +371,7 @@ export const buildJobExternalVerificationsObject = (
       error: verifierData.error,
       statusUrl,
       explorerUrl,
+      contractApiUrl,
     };
 
     switch (verifier) {
@@ -508,7 +518,7 @@ export class EtherscanVerifyApiService implements WStorageService {
     } catch (error: any) {
       // Store the external verification result if we have job data
       if (jobData) {
-        this.storeExternalVerificationResult(jobData, {
+        await this.storeExternalVerificationResult(jobData, {
           error: error.message,
         });
       }
@@ -530,14 +540,16 @@ export class EtherscanVerifyApiService implements WStorageService {
       return;
     }
 
-    // Handle the "already verified" case for Blockscout by storing
-    // { verificationId: "BLOCKSCOUT_ALREADY_VERIFIED" }
+    // Handle the "already verified" case for Blockscout and Etherscan by storing
+    // { verificationId: "VERIFIER_ALREADY_VERIFIED" }
     if (
-      this.IDENTIFIER === WStorageIdentifiers.BlockscoutVerify &&
-      response.result === "Smart-contract already verified."
+      (this.IDENTIFIER === WStorageIdentifiers.BlockscoutVerify &&
+        response.result === "Smart-contract already verified.") ||
+      (this.IDENTIFIER === WStorageIdentifiers.EtherscanVerify &&
+        response.result === "Contract source code already verified")
     ) {
-      this.storeExternalVerificationResult(jobData, {
-        verificationId: BLOCKSCOUT_ALREADY_VERIFIED,
+      await this.storeExternalVerificationResult(jobData, {
+        verificationId: VERIFIER_ALREADY_VERIFIED,
       });
       return;
     }
@@ -549,14 +561,14 @@ export class EtherscanVerifyApiService implements WStorageService {
         verificationJobId: jobData.verificationId,
         error: response.result,
       });
-      this.storeExternalVerificationResult(jobData, {
+      await this.storeExternalVerificationResult(jobData, {
         error: response.result,
       });
       return;
     }
 
     // Record the result of the successful submission
-    this.storeExternalVerificationResult(jobData, {
+    await this.storeExternalVerificationResult(jobData, {
       verificationId: response.result,
     });
   }
@@ -634,7 +646,7 @@ export class EtherscanVerifyApiService implements WStorageService {
     if (!response.ok) {
       const responseText = await response.text();
       if (jobData) {
-        this.storeExternalVerificationResult(jobData, {
+        await this.storeExternalVerificationResult(jobData, {
           error: responseText,
         });
       }
@@ -654,11 +666,11 @@ export class EtherscanVerifyApiService implements WStorageService {
       // Blockscout does not return a verification ID,
       // so we use a fixed string to indicate a successful submission
       if (res.message === "Smart-contract verification started") {
-        this.storeExternalVerificationResult(jobData, {
+        await this.storeExternalVerificationResult(jobData, {
           verificationId: "BLOCKSCOUT_VYPER_SUBMITTED",
         });
       } else {
-        this.storeExternalVerificationResult(jobData, {
+        await this.storeExternalVerificationResult(jobData, {
           error: res.message,
         });
       }