+++ services/server/test/integration/apiv2/jobs.spec.ts
@@ -116,7 +116,7 @@ describe("GET /v2/verify/:verificationId", function () {
           verificationId: "some-external-id",
         },
         BlockscoutVerify: {
-          verificationId: "BLOCKSCOUT_ALREADY_VERIFIED",
+          verificationId: "VERIFIER_ALREADY_VERIFIED",
         },
         RoutescanVerify: {
           error: "some error",
@@ -130,9 +130,12 @@ describe("GET /v2/verify/:verificationId", function () {
           explorerUrl:
             "https://etherscan.io/address/" +
             chainFixture.defaultContractAddress,
+          contractApiUrl:
+            "https://api.etherscan.io/api?module=contract&action=getabi&chainid=31337&address=" +
+            chainFixture.defaultContractAddress,
         },
         blockscout: {
-          verificationId: "BLOCKSCOUT_ALREADY_VERIFIED",
+          verificationId: "VERIFIER_ALREADY_VERIFIED",
           explorerUrl:
             "https://eth.blockscout.io/address/" +
             chainFixture.defaultContractAddress,